@@ -29,11 +29,11 @@
     "mysql": "^2.15.0"
   },
   "devDependencies": {
-    "@types/node": "^8.5.2",
+    "@types/node": "^8.5.8",
     "concurrently": "^3.5.1",
-    "eslint": "^4.14.0",
+    "eslint": "^4.15.0",
     "faker": "^4.1.0",
-    "nodemon": "^1.14.3",
+    "nodemon": "^1.14.9",
     "typescript": "^2.6.2"
   }
 }
