@@ -1,43 +1,30 @@
-import { getRandomData, controlLimitSeeder, compareUnique } from '../helpers/index';
+import {
+  randomItem
+} from '../helpers/index';
 import User from '../models/user';
 import Comment from '../models/comment';
 
-let limit = 60;
-export const seed = function(knex, Promise) {
-    // Deletes ALL existing entries
-    return knex('likes_comments').del()
-        .then(async() => {
-            const faker = require('faker/locale/es');
-            // Inserts seed entries
-            const datas = [];
-            const commentDatas = await Comment.get();
-            const userDatas = await User.get();
-
-            const errorLimitInsert = controlLimitSeeder([commentDatas, userDatas], limit);
-            if (errorLimitInsert.TheDataIsNotCorrect) {
-                if (errorLimitInsert.IsMissingDataLength) {
-                    throw new Error(`Tabla 'likes_comments' Data` + `\x1b[31mInsuficientes FK\x1b[0m\n`);
-                } else if (errorLimitInsert.IsNotMinimumData) {
-                    console.log(`Tabla 'likes_comments' change Limit`, `\x1b[33m${errorLimitInsert.limitMax}\x1b[0m`);
-                    limit = errorLimitInsert.limitMax;
-                }
-            }
-            for (let index = 0; index < limit; index++) {
-
-                let data = {};
-                do {
-                    data = {
-                        comment_id: (await getRandomData(commentDatas)).attributes.id,
-                        user_id: (await getRandomData(userDatas)).attributes.id,
-                    };
-                } while (compareUnique(datas, data, ['comment_id', 'user_id']));
-                datas.push(data);
-            }
-
-            return knex('likes_comments').insert(datas)
-                .then(console.log(`Tabla 'likes_comments' Datos`, '\x1b[32mOK\x1b[0m'))
-                .catch((error) => {
-                    console.log(`Tabla 'likes_comments' Datos`, `\x1b[31mFAIL\x1b[0m\n${error}`)
-                });
-        });
+const limit = 20;
+export const seed = (knex) => {
+  return knex('likes_comments').del().then(async() => {
+    const faker = require('faker/locale/es');
+    const items = [];
+    const comments = (await Comment.get()).toJSON().map(item => item.id);
+    const users = (await User.get()).toJSON().map(item => item.id);
+    let comment_user;
+    for (let index = 0; index < limit; index++) {
+      do {
+        comment_user = {
+          comment_id: randomItem(comments),
+          user_id: randomItem(users),
+        };
+      } while (items.findIndex(item =>
+          item.comment_id === comment_user.comment_id && item.user_id && comment_user.user_id
+        ) !== -1);
+      items.push(comment_user);
+    }
+    return knex('likes_comments').insert(items)
+      .then(console.log(`Tabla 'likes_comments' Datos`, '\x1b[32mOK\x1b[0m'))
+      .catch((error) => console.log(`Tabla 'likes_comments' Datos`, `\x1b[31mFAIL\x1b[0m\n${error}`));
+  });
 };
